DF-12948 Fps to frametime
Former-commit-id: d7ab3d33cf5f53373f2cc26ac10efdacd1866c32

+++ Programs/OverdrawPerformanceTester/Classes/ViewSceneScreen.h
@@ -25,7 +25,7 @@ protected:
 public:
     struct FrameData
     {
-        DAVA::int32 FPS;
+        DAVA::float32 FPS;
         DAVA::float32 Overdraw;
     };
